feat: clean project: fix paths, rename routes files, invalid span use
