+++ src/router/index.js
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import { CUSTOMER } from '../constants';
+import { CUSTOMER } from '@/constants';
 import routes from './routes';
 const router = createRouter({
   history: createWebHistory(),
